common max label
